Use one libddwaf.so for both glibc and musl
This is achived by compiling against musl (dynamically) and against
libc++ (statically). libc++ is also built against musl. For good
measure, compile against compiler-rt and libunwind statically (though
compiling statically against gcc's equivalents would also probably
work). libc++ is also compiled against compiler-rt and libunwind.

The compiler used for building both libc++ and libddwaf is clang 11.1
(as provided by the alpine packages). The alpine
compiler-rt/compiler-rt-static packages are also used. libc++
11.1/libwind 11.1 are built in custom package, whose APKBUILD is
provided.

The resulting libddwaf.so is then post-processed to removed the NEEDED
entry for the musl DSO, so that it can be run in glibc systems. The
assumption here being that the dynamic linker will provide the imported
libc symbols without the NEEDED elf entry. Another assumption is that
the build against musl will be binary compatible with glibc.

We still have two JNI binaries, onde for glibc and one for musl. Note
that many symbols that the musl provides on its DSO, glibc provides in
libm, libdl, libpthread and, for old versions, librt. Consequently, the
glibc JNI library MUST request that those libraries be loaded (with
NEEDED entries) so that they become available for the universal
libddwaf.so.